requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==1.1.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2>=3.1.2,<4.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv>=0.10.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug>=2.0.0
WTForms==3.0.1
waitress==2.1.2